% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barfit.R
\name{barfit}
\alias{barfit}
\title{Estimating a buffered autoregressive model with Least Squares estimation.}
\usage{
barfit(y, z = y, d = 1, p0 = 1, p1 = 1, r = c(0.1, 0.9), search = "quantile")
}
\arguments{
\item{y}{A numeric vector representing the outcome time series.}

\item{z}{A numeric vector representing the threshold time series.}

\item{d}{A vector with one or more values for the delay parameter.}

\item{p0}{The order of regime 0.}

\item{p1}{The order of regime 1.}

\item{r}{A numeric vector of length 2 with (boundary) threshold values}

\item{search}{What type of boundary values of r are specified to find
optimal threshold values: "quantile" for quantile values, "scale" for
values on the scale of z, "none" if r is fixed}
}
\value{
An object of class bar.
}
\description{
Estimating a buffered autoregressive model with Least Squares estimation.
}
\examples{
a <- 1
}
