% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hystar-package.R
\docType{package}
\name{hystar-package}
\alias{hystar}
\alias{hystar-package}
\title{hystar: Simulation and Estimation of the Hysteretic TAR Model}
\description{
Simulates data from and estimates parameters of the hysteretic threshold autoregressive (HysTAR) model. Estimation is implemented using the conditional least squares method. See Li, Guan, Li and Yu (2015) \doi{10.1093/biomet/asv017} for details about the model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/daandejongen/hystar/}
  \item Report bugs at \url{https://github.com/daandejongen/hystar/issues/}
}

}
\author{
\strong{Maintainer}: Daan de Jong \email{d.dejong@uu.nl} (\href{https://orcid.org/0000-0001-6034-0136}{ORCID})

Other contributors:
\itemize{
  \item European Research Council [funder]
}

}
\keyword{internal}
