% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hystar_fit.R
\name{hystar_fit}
\alias{hystar_fit}
\title{Estimate the hysTAR model using conditional least squares estimation}
\usage{
hystar_fit(
  y,
  z,
  d = 1L,
  p0 = 1L,
  p1 = 1L,
  p_select = "bic",
  r = c(0.1, 0.9),
  thin = TRUE
)
}
\arguments{
\item{y}{A numeric vector representing the outcome time series. Can be simulated
with \code{\link[=hystar_sim]{hystar_sim()}}.}

\item{z}{A numeric vector representing the threshold time series.}

\item{d}{A numeric vector with one or more values for the search space
of the delay parameter. Defaults to 1.}

\item{p0}{A numeric vector with one or more values for the search space
of the autoregressive order of Regime 0. Defaults to 1.}

\item{p1}{Same as \code{p0}, but for regime 1.}

\item{p_select}{Which information criterion should be minimized to select
the orders \eqn{p_0} and \eqn{p_1} Choices:
* "aic" (Akaike Information Criterion)
* "aicc" (Corrected Akaike Information Criterion)
* "bic" (default, Bayesian Information Criterion)}

\item{r}{A vector or a matrix with search values for \eqn{\hat{r}_0, \hat{r}_1}.
Defaults to \code{c(.1, .9)}.
* If \code{r} is a vector, its length must be 2, such that it represents two quantiles
within which the threshold value should be searched.
* If \code{r} is a matrix, it must have two columns,
such that each row represents a pair \eqn{r_0 \le r_1} to test.
You can use a matrix with one row if you don't want to estimate
the thresholds. Note that the values in these matrix are not quantiles,
but directly values on the scale of \code{z}.}

\item{thin}{Logical, only relevant when \code{r} is a vector.
If \code{TRUE} (default), the search values for \eqn{r} are
\code{quantile(z, seq(r[1], r[2], by = .01))}. This drastically improves the speed
without much cost of estimation precision. Note that this is a purely practical
choice with no theoretical justification (yet).
If \code{FALSE}, all observed unique values of \code{z} between
\code{quantile(z, r[1])} and \code{quantile(z, r[2])} will be considered.}
}
\value{
An object of S3 class \code{hystar_fit}, which is a \code{list} containing the following
items:
\itemize{
\item \verb{$data}. A \code{data.frame} of class \code{hystar_data}, containing
\itemize{
\item \code{y}, the outcome variable
\item \code{z}, the threshold variable
\item \code{H}, a logical vector that indicates at which time points the hysteresis
effect is happening. Note that this vector starts with \code{NA}(s), since the first
\eqn{\max\{p_0, p_1, d\}} values are not predicted in the HysTAR model.
\item \code{R}, the regime indicator vector. (Also starts with \code{NA}(s).)
}
\item \verb{$residuals}. Also accessible with the \code{residuals()} S3 method.
\item \verb{$coefficients}, a named vector with the estimated coefficients.
With the \code{coef()} S3 method, the coefficients are represented in a matrix.
Use the \code{confint()} method to get the confidence intervals of the estimates.
\item \verb{$delay}, a named scalar with the estimate for the delay parameter.
\item \verb{$thresholds}, a named vector with the estimates of the tresholds.
\item \verb{$orders}, a named vector with the estimates of the orders.
\item \verb{$resvar}, a named vector with the estimates of the residual variances.
\item \verb{$rss}, the minimized residual sum of squares.
\item \verb{$ic}, a named vector with the aic, the corrected aic and the bic.
\item \verb{$n}, a named vector with the total effective observations and the
effective obeservations in regime 0 and regime 1.
\item \verb{$eff}, a vector with the time indicators of the effective observations.
\item \verb{$equiv}, a matrix containing equivalent estimates for the delay and thresholds,
i.e., estimates that imply exactly the same regime indicator vector, and
as a result the same minimal residual sum of squares.
}

Implemented generics for the \code{hystar_fit} class:
\itemize{
\item \code{plot()} plots the \code{z} variable and the \code{y} variable above one another.
Shading of the background visualizes the regimes. Thresholds are drawn as
horizontal lines in the \code{z} plot.
\item \code{summary()}, this also provides the p-values and standard errors for the
estimates of the coefficients.
\item \code{print()} prints the estimates within the mathematical representation of the model.
Note that the scalar multiplied with \code{e[t]} is the standard deviation
of the residuals, \emph{not} the variance. See also the model definition above.
\item \code{coef()}
\item \code{confint()}
\item \code{residuals()}
\item \code{fitted()}
\item \code{nobs()}
}
}
\description{
This function allows you to estimate the parameters of the hysteretic threshold
autoregressive (hysTAR) model.
}
\section{The hysTAR model}{
The hysTAR model is defined as:

\eqn{ y_t = \begin{cases} \phi_0^{(0)} + \phi_1^{(0)} y_{t-1} + \cdots +
  \phi_{p_0}^{(0)} y_{t-p_0} + \sigma_{(0)} \varepsilon_t & \text{if}~R_{t} = 0
  \\ \phi_0^{(1)} + \phi_1^{(1)} y_{t-1} + \cdots + \phi_{p_1}^{(1)} y_{t-p_1}
  + \sigma_{(1)} \varepsilon_t & \text{if}~R_{t} = 1, \\ \end{cases} }

with \eqn{ R_t = \begin{cases} 0 & \mathrm{if} \, z_{t-d} \in (-\infty,
  r_{0}] \\ R_{t-1} & \mathrm{if} \, z_{t-d} \in (r_0, r_1] \\ 1 & \mathrm{if}
  \, z_{t-d} \in (r_1, \infty), \\ \end{cases} }

where \eqn{p_i} denotes the order of regime \eqn{i \in \{0,1\}} with
coefficients \eqn{\phi_0^{(i)}, \dots, \phi_{p_i}^{(i)} \in (-1, 1)},
\eqn{\sigma_{(i)}} is the standard deviation of the residuals, and \eqn{d \in
  \{1, 2, \dots\}} is a delay parameter. The parameters of primary interest are
the thresholds \eqn{r_0 \le r_1}.
}

\examples{
z <- z_sim(n_t = 200, n_switches = 5, start_regime = 1)
sim <- hystar_sim(z = z, r = c(-.5, .5), d = 2, phi_R0 = c(0, .6), phi_R1 = 1,
resvar = c(1, 1))
plot(sim)
fit <- hystar_fit(y = sim$data$y, z = z)
summary(fit)
}
\references{
Li, Guodong, Bo Guan, Wai Keung Li, en Philip L. H. Yu.
‘Hysteretic Autoregressive Time Series Models’. Biometrika 102, nr. 3
(september 2015): 717–23.
}
\author{
Daan de Jong.
}
