% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{hystar_info}
\alias{hystar_info}
\title{Get more information about the hystar package}
\usage{
hystar_info()
}
\value{
Nothing
}
\description{
directs you to the hystar website
https://daandejongen.github.io/hystar/index.html
}
