% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.hystar_fit}
\alias{plot.hystar_fit}
\title{Plot method of the \code{hystar_fit} class}
\usage{
plot(hystar_fit_object)
}
\arguments{
\item{x}{an object of class \code{hystar_fit}}

\item{y}{not needed}

\item{main}{title of the plot}

\item{...}{further arguments passed to \code{plot()}, see Details.}
}
\description{
Plots the time series of the control variable \code{z} and
the outcome variable \code{y} on top of each other, with background colors
indicating the regimes.
}
\details{
You can specify the following additional parameters (with defaults):
Names:
\itemize{
\item \code{name_regime0 = "Regime 0"}
\item \code{name_regime1 = "Regime 1"}
\item \code{name_regime_unknown = "Unknown"}
\item \code{zlab = "z"}, label of the y-axis in the z-plot
\item \code{ylab = "y"}, label of the y-axis in the y-plot
\item \code{xlab = "x"}, label of the x-axis
}

Colors:
\itemize{
\item \code{color_regime0 = "white"}
\item \code{color_regime0 = "grey85"}
\item \code{color_regime_unknown = "AA001137"}
\item \code{color_z = "grey25"}
\item \code{color_y = "grey25"}
}

Lines:
\code{line_type_z = 1}
\code{line_width_z = 2.5}
\code{line_type_y = 1}
\code{line_width_y = 2.5}

Other:
\code{show_legend = TRUE}
}
