% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{z_sim}
\alias{z_sim}
\title{Simulate the threshold/control variable Z}
\usage{
z_sim(start_regime, n_switches, n_t, form = "cosine", scale = c(-1, 1))
}
\arguments{
\item{start_regime}{a}

\item{n_switches}{a}

\item{n_t}{a}

\item{form}{a}

\item{scale}{a}
}
\value{
a
}
\description{
blabla
}
\details{
The number of switches \code{n_switches} can differ ultimately based on r and d.
and if starting regime is weird. (only for sine function)
}
\examples{
1
}
