% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar.R
\name{bar}
\alias{bar}
\title{Estimating a buffered autoregressive model with Least Squares estimation.}
\usage{
bar(y, z = y, d, p, r, search)
}
\arguments{
\item{y}{A numeric vector representing the outcome time series.}

\item{z}{A numeric vector representing the threshold time series.}

\item{d}{A vector with one or more values for the delay parameter.}

\item{r}{A numeric vector of length 2 with (boundary) threshold values}

\item{search}{}

\item{p0}{A vector with one or more values for the order of regime 0.}

\item{p1}{A vector with one or more values for the order of regime 1.}
}
\value{
An object of class bar.
}
\description{
Estimating a buffered autoregressive model with Least Squares estimation.
}
\examples{
a <- 1
}
