% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar.R
\name{bar}
\alias{bar}
\title{Create a proportion chart}
\usage{
bar(data, colors, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{a \code{list} with items \code{"name"} and \code{"value"}
or a \code{data.frame} with labels in first column and values in second.}

\item{colors}{Named \code{list} with names corresponding to labels and values to colors.}

\item{width, height}{Width and height for the chart.}

\item{elementId}{Explicit ID for the widget.}
}
\description{
Create a proportion chart
}
