% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barsim.R
\name{barsim}
\alias{barsim}
\title{Title}
\usage{
barsim(
  r,
  d,
  phi,
  psi,
  resvar,
  init_vals = NULL,
  start_regime = 0,
  z = NULL,
  length = NULL,
  n_switches = NULL
)
}
\arguments{
\item{r}{A vector of length 2 containing the trhesholds}

\item{d}{The delay parameter}

\item{phi}{A vector containing the autoregressive coefficients of regime 0.
Note that a constant/intercept must be specified. If you don't want a
constant, you can just set it to zero.}

\item{psi}{A vector containing the autoregressive coefficients of regime 0}

\item{resvar}{A vector containing the residual variances of regime 0 and 1}

\item{init_vals}{Initial values: The first length(z)-max(p0, p1, d) values
are fixed in the model.}

\item{start_regime}{A starting regime, 0 or 1}

\item{z}{A numeric vector with observations of the threshold/control variable}
}
\value{
Simulated time series generated by the BAR model
}
\description{
Title
}
