% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hystar-package.R
\docType{package}
\name{hystar-package}
\alias{hystar}
\alias{hystar-package}
\title{hystar: Fit the Hysteretic Threshold Autoregressive Model}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Estimate parameters of the hysteretic threshold autoregressive (HysTAR) model, using conditional least squares. In addition, you can generate time series data from the HysTAR model. For details, see Li, Guan, Li and Yu (2015) \doi{10.1093/biomet/asv017}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/daandejongen/hystar/}
  \item \url{https://daandejongen.github.io/hystar/}
  \item Report bugs at \url{https://github.com/daandejongen/hystar/issues/}
}

}
\author{
\strong{Maintainer}: Daan de Jong \email{daandejong94@gmail.com} (\href{https://orcid.org/0000-0001-6034-0136}{ORCID})

Other contributors:
\itemize{
  \item Utrecht University [copyright holder]
  \item European Research Council [funder]
}

}
\keyword{internal}
