% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hystar-package.R
\docType{package}
\name{hystar-package}
\alias{hystar}
\alias{hystar-package}
\title{hystar: Simulation And Least Squares Estimation of the hysteretic TAR Model}
\description{
The hysteretic threshold autoregressive (hysTAR) model describes a time series that follows one of two possible AR processes according to a hysteretic switching mechanism. With 'hystar', you can simulate data from this model and estimate parameter values based on empirical or simulated data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/daandejongen/bar/}
  \item Report bugs at \url{https://github.com/daandejongen/bar/issues/}
}

}
\author{
\strong{Maintainer}: Daan de Jong \email{d.dejong@uu.nl} (\href{https://orcid.org/0000-0001-6034-0136}{ORCID})

}
\keyword{internal}
