% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hystar_fit.R
\name{hystar_fit}
\alias{hystar_fit}
\title{Estimating the hysTAR model using least squares estimation}
\usage{
hystar_fit(
  y,
  z = y,
  d = 1,
  p0 = 1,
  p1 = p1,
  r = c(0.1, 0.9),
  r_type = "quantile",
  r_select = "smallest"
)
}
\arguments{
\item{y}{A numeric vector representing the outcome time series.}

\item{z}{A numeric vector representing the threshold time series.
If \code{z} = \code{y} (default), the model is a hysSETAR
(hysteretic self-exciting threshold autoregressive) model.}

\item{d}{A numeric vector with one or more values for the search space
of the delay parameter.}

\item{p0}{A numeric vector with one or more values for the search space
of the autoregressive order of Regime 0.}

\item{p1}{A numeric vector with one or more values for the search space
of the autoregressive order of Regime 1. By default equal to p0.}

\item{r}{A vector or a matrix. If \code{r} is a vector, its length must be 2,
such that it represents the interval in which the threshold value
should be searched. If \code{r} is a matrix, it must have two columns,
such that each row represents a pair \eqn{r_0 \le r_1} to test.
You can use a matrix with one row if you don't want to estimate
the thresholds.}

\item{r_type}{What type of values are gievn in \code{r}.
\code{"quantile"} for quantile values and
\code{"scale"} for values on the scale of \code{z}.}

\item{r_select}{When multiple pairs of thresholds yield the same residual
sums of squares, which one should be selected?
Options: \code{"widest"} or \code{"smallest"} (default).}
}
\value{
An object of class bar.
}
\description{
The hysTAR model
}
\examples{
model <- barfit(y = y, z = z, d = c(1, 3), p0 = 1:4, p1 = 2)
summary(model)
plot(model$data)
}
