% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hystar_fit.R
\name{hystar_fit}
\alias{hystar_fit}
\title{Estimating the hysTAR model using least squares estimation}
\usage{
hystar_fit(
  y,
  z = y,
  d = 1,
  p0 = 1,
  p1 = p1,
  r = c(0.1, 0.9),
  r_type = "quantile",
  r_select = "smallest"
)
}
\arguments{
\item{y}{A numeric vector representing the outcome time series.}

\item{z}{A numeric vector representing the threshold time series.
If \code{z} = \code{y} (default), the model is a hysSETAR
(hysteretic self-exciting threshold autoregressive) model.}

\item{d}{A numeric vector with one or more values for the search space
of the delay parameter. Defaults to 1.}

\item{p0}{A numeric vector with one or more values for the search space
of the autoregressive order of Regime 0. Defaults to 1.}

\item{p1}{A numeric vector with one or more values for the search space
of the autoregressive order of Regime 1. By default equal to p0.
Defaults to \code{p0}.}

\item{r}{A vector or a matrix. If \code{r} is a vector, its length must be 2,
such that it represents the interval in which the threshold value
should be searched. If \code{r} is a matrix, it must have two columns,
such that each row represents a pair \eqn{r_0 \le r_1} to test.
You can use a matrix with one row if you don't want to estimate
the thresholds. Defaults to \eqn{(0.1, 0.9)}.}

\item{r_type}{What type of values are given in \code{r}:
\itemize{
\item \code{"quantile"} (default) for quantile values and
\item \code{"scale"} for values on the scale of \code{z}.
}}

\item{r_select}{When multiple pairs of thresholds yield the same residual
sums of squares, which one should be selected? Options:
\itemize{
\item \code{"smallest"} (default): \eqn{\mathrm{argmin}_{r_0, r_1} \{|r_0 - r_1|\} },
\item \code{"widest"}: \eqn{\mathrm{argmax}_{r_0, r_1} \{|r_0 - r_1|\} }.
}}
}
\value{
An object of class \link{bar}.
}
\description{
The hysTAR model is defined as:

\eqn{
    y_t =
    \begin{cases}
    \phi_0^{(0)} + \phi_1^{(0)} y_{t-1} + \cdots + \phi_{p_0}^{(0)}
    y_{t-p_0} + \sigma_{(0)} \varepsilon_t & \text{if}~R_{t} = 0 \\
    \phi_0^{(1)} + \phi_1^{(1)} y_{t-1} + \cdots + \phi_{p_1}^{(1)}
    y_{t-p_1} + \sigma_{(1)} \varepsilon_t & \text{if}~R_{t} = 1, \\
    \end{cases}
}

where
\eqn{
    R_t = \begin{cases}
    0 & \mathrm{if} \, z_{t-d} \in (-\infty, r_{0}] \\
    R_{t-1} & \mathrm{if} \, z_{t-d} \in (r_0, r_1] \\
    1 & \mathrm{if} \, z_{t-d} \in (r_1, \infty), \\
    \end{cases}
}
}
\details{
Coefficient estimates are offered in various ways:
as the element of the \code{hystar} (list) object: a named vector.
in \code{coef(hystar)}: as a 2 x max(order)+1 matrix, only estimates.
in \code{summary(hystar)}: a matrix with the SEs and p-values.
in \code{print(hystar)}: in the model equations.
}
\examples{
model <- barfit(y = y, z = z, d = c(1, 3), p0 = 1:4, p1 = 2)
summary(model)
plot(model$data)
}
