% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hystar_fit.R
\name{hystar_fit}
\alias{hystar_fit}
\title{Estimating the hysTAR model using least squares estimation}
\usage{
hystar_fit(
  y,
  z,
  d = 1L,
  p0 = 1L,
  p1 = 1L,
  p_select = "bic",
  r = c(0.1, 0.9),
  thin = TRUE
)
}
\arguments{
\item{y}{A numeric vector representing the outcome time series.}

\item{z}{A numeric vector representing the threshold time series.
If \code{z} = \code{y}, the model is a hysSETAR
(hysteretic self-exciting threshold autoregressive) model.}

\item{d}{A numeric vector with one or more values for the search space
of the delay parameter. Defaults to 1.}

\item{p0}{A numeric vector with one or more values for the search space
of the autoregressive order of Regime 0. Defaults to 1.}

\item{p1}{Same as \code{p0}, but for regime 1.}

\item{p_select}{Which information criterion should be minimized to select
the orders p0 and p1. Choices:
\itemize{
\item "aic" (Akaike Information Criterion)
\item "aicc" (Corrected Akaike Information Criterion)
\item "bic" (default, Bayesian Information Criterion)
}}

\item{r}{A vector or a matrix with search values for \eqn{r_0, r_1}.
Defaults to \code{c(.1, .9)}.
\itemize{
\item If \code{r} is a vector, its length must be 2, such that it represents two quantiles
within which the threshold value should be searched.
\item If \code{r} is a matrix, it must have two columns,
such that each row represents a pair \eqn{r_0 \le r_1} to test.
You can use a matrix with one row if you don't want to estimate
the thresholds. Note that the values in these matrix are not quantiles,
but directly values on the scale of \code{z}.
}}

\item{thin}{Logical, only relevant when \code{r} is a vector.
If \code{TRUE} (default), the search values for \eqn{r} are
\code{quantile(z, seq(r[1], r[2], by = .01))}.
This drastically improves the speed without much
cost of estimation precision.
Note that this is a purely practical choice
with no theoretical justification (yet).
If \code{FALSE}, all observed unique values of \code{z} between
\code{quantile(z, r[1])} and \code{quantile(z, r[2])} will be considered.}
}
\value{
An object of S3 class \code{hystar}, which is a list containing the following
items:
\itemize{
\item \verb{$data}. A \code{data.frame} of class \code{hystar_data}, containing
\itemize{
\item \code{y}, the outcome variable
\item \code{z}, the threshold variable
\item \code{H}, a logical vector that indicates at which time points the hysteresis
effect is happening. Note that this vector starts with \code{NA}(s), since the first
\eqn{\max\{p_0, p_1, d\}} values are not predicted in the HysTAR model.
\item \code{R}, the regime indicator vector. (Also starts with \code{NA}(s).)
}
\item \verb{$residuals}. Also accessible with the \code{residuals()} S3 method.
\item \verb{$coefficients}, a named vector with the estimated coefficients.
With the \code{residuals()} S3 method, the residuals are represented in a matrix.
Use the \code{confint()} method to get the confidence intervals of the estimates.
\item \verb{$delay}, a named scalar with the estimate for the delay parameter.
\item \verb{$thresholds}, a named vector with the estimates of the tresholds.
\item \verb{$orders}, a named vector with the estimates of the orders.
\item \verb{$resvar}, a named vector with the estimates of the residual variances.
\item \verb{$rss}, the minimized residual sum of squares.
\item \verb{$ic}, a named vector with the aic, the corrected aic and the bic.
\item \verb{$n}, a named vector with the total effective observations and the
effective obeservations in regime 0 and regime 1.
\item \verb{$eff}, a vector with the time indicators of the effective observations.
\item \verb{$equiv}, a matrix containing equivalent estimates for \eqn{d, p_0} and \eqn{p_1},
i.e., estimates that imply exactly the same regime indicator vector, and
as a result the same minimal residual sum of squares.
}

S3 methods:
\itemize{
\item The \code{hystar_data} class has a \code{plot()} method.
\item \code{summary()}, this also provides the p-values and standard errors for the
estimates of
\eqn{\phi_0^{(0)}, \dots, \phi_{p_0}^{(0)}, \phi_0^{(1)}, \dots, \phi_{p_1}^{(1)}}.
\item \code{print()}
\item \code{coef()}
\item \code{confint()}
\item \code{residuals()}
\item \code{fitted()}
\item \code{nobs()}
}
}
\description{
The hysTAR model is defined as:

\eqn{
    y_t =
    \begin{cases}
    \phi_0^{(0)} + \phi_1^{(0)} y_{t-1} + \cdots + \phi_{p_0}^{(0)}
    y_{t-p_0} + \sigma_{(0)} \varepsilon_t & \text{if}~R_{t} = 0 \\
    \phi_0^{(1)} + \phi_1^{(1)} y_{t-1} + \cdots + \phi_{p_1}^{(1)}
    y_{t-p_1} + \sigma_{(1)} \varepsilon_t & \text{if}~R_{t} = 1, \\
    \end{cases}
}

where
\eqn{
    R_t = \begin{cases}
    0 & \mathrm{if} \, z_{t-d} \in (-\infty, r_{0}] \\
    R_{t-1} & \mathrm{if} \, z_{t-d} \in (r_0, r_1] \\
    1 & \mathrm{if} \, z_{t-d} \in (r_1, \infty), \\
    \end{cases}
}
}
\details{
Coefficient estimates are offered in various ways:
as the element of the \code{hystar} (list) object: a named vector.
in \code{coef(hystar)}: as a 2 x max(order)+1 matrix, only estimates.
in \code{summary(hystar)}: a matrix with the SEs and p-values.
in \code{print(hystar)}: in the model equations.
}
\examples{
model <- hystar_fit(y = y, z = z, d = c(1, 3), p0 = 1:4, p1 = 2)
summary(model)
plot(model$data)
}
