% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hystar_sim.R
\name{hystar_sim}
\alias{hystar_sim}
\title{Simulating data from the hysTAR model}
\usage{
hystar_sim(
  r,
  d,
  phi_R0,
  phi_R1,
  resvar,
  init_vals = NULL,
  z = NULL,
  n_t = NULL,
  n_switches = NULL,
  start_regime = NULL
)
}
\arguments{
\item{r}{A vector of length 2, representing the threshold values \eqn{r_0} and \eqn{r_1}.}

\item{d}{A number in \eqn{\{1, 2, \dots\}} representing the value of the delay parameter.}

\item{phi_R0}{A vector containing the constant and autoregressive parameters
\eqn{(\phi_0^{(0)}, \phi_1^{(0)}, \dots, \phi_{p_0}^{(0)})} of Regime 0.
Note that the first value of this vector is always interpreted as the constant.}

\item{phi_R1}{The same as \code{phi_R0}, but for Regime 1.}

\item{resvar}{A numeric vector of length 2 representing the residual variances
\eqn{\sigma_{(0)}^2} and \eqn{\sigma_{(1)}^2}.}

\item{init_vals}{Optionally, a vector of length \eqn{\max{d, p_0, p_1}} to provide the
first values of the simulated \eqn{y} variable. See Details. When omitted, the first values
are simulated from \eqn{\mathcal{N(\mu_i, \sigma_{(i)}^2)}}, where
\eqn{\mu_y = \mathbb{E}(Y_t|R_t=i) = \frac{\phi_0^{(0)}}{1 - \phi_1^{(0)} - \cdots - \phi_{p_i}^{(0)}}}
and \eqn{i} is the starting regime.}

\item{start_regime}{}
}
\description{
The hysTAR model is defined as:

\eqn{
    y_t =
    \begin{cases}
    \phi_0^{(0)} + \phi_1^{(0)} y_{t-1} + \cdots + \phi_{p_0}^{(0)}
    y_{t-p_0} + \sigma_{(0)} \varepsilon_t & \text{if}~R_{t} = 0 \\
    \phi_0^{(1)} + \phi_1^{(1)} y_{t-1} + \cdots + \phi_{p_1}^{(1)}
    y_{t-p_1} + \sigma_{(1)} \varepsilon_t & \text{if}~R_{t} = 1, \\
    \end{cases}
}

where
\eqn{
    R_t = \begin{cases}
    0 & \mathrm{if} \, z_{t-d} \in (-\infty, r_{0}] \\
    R_{t-1} & \mathrm{if} \, z_{t-d} \in (r_0, r_1] \\
    1 & \mathrm{if} \, z_{t-d} \in (r_1, \infty), \\
    \end{cases}
}
}
